Move and rename generator files so Zeitwerk stops exploding
