+++ lib/shiny/plugin/plugin_generator.rb
@@ -19,7 +19,7 @@
 require_relative 'plugin_generator_base'
 
 module Shiny
-  module Generators
+  module Plugin
     # ShinyCMS plugin generator
     class PluginGenerator < PluginGeneratorBase # :nodoc:
       alias plugin_path app_path
